feat(extras): big rework of default extras
