+++ lua/lazyvim/plugins/extras/editor/fzf.lua
@@ -286,9 +286,6 @@ return {
   {
     "neovim/nvim-lspconfig",
     opts = function()
-      if LazyVim.pick.want() ~= "fzf" then
-        return
-      end
       local Keys = require("lazyvim.plugins.lsp.keymaps").get()
       -- stylua: ignore
       vim.list_extend(Keys, {
