+++ lua/lazyvim/plugins/extras/editor/snacks_picker.lua
@@ -136,9 +136,6 @@ return {
   {
     "neovim/nvim-lspconfig",
     opts = function()
-      if LazyVim.pick.want() ~= "snacks" then
-        return
-      end
       local Keys = require("lazyvim.plugins.lsp.keymaps").get()
       -- stylua: ignore
       vim.list_extend(Keys, {
