+++ lua/lazyvim/plugins/extras/editor/telescope.lua
@@ -61,9 +61,6 @@ return {
   {
     "nvim-telescope/telescope.nvim",
     cmd = "Telescope",
-    enabled = function()
-      return LazyVim.pick.want() == "telescope"
-    end,
     version = false, -- telescope did only one release, so use HEAD for now
     dependencies = {
       {
@@ -267,9 +264,6 @@ return {
   {
     "stevearc/dressing.nvim",
     lazy = true,
-    enabled = function()
-      return LazyVim.pick.want() == "telescope"
-    end,
     init = function()
       ---@diagnostic disable-next-line: duplicate-set-field
       vim.ui.select = function(...)
@@ -287,9 +281,6 @@ return {
   {
     "neovim/nvim-lspconfig",
     opts = function()
-      if LazyVim.pick.want() ~= "telescope" then
-        return
-      end
       local Keys = require("lazyvim.plugins.lsp.keymaps").get()
       -- stylua: ignore
       vim.list_extend(Keys, {
