+++ lua/lazyvim/plugins/extras/lang/sql.lua
@@ -48,7 +48,7 @@ return {
       vim.api.nvim_create_autocmd("FileType", {
         pattern = sql_ft,
         callback = function()
-          if LazyVim.cmp_engine() == "nvim-cmp" then
+          if LazyVim.has_extra("coding.nvim-cmp") then
             local cmp = require("cmp")
 
             -- global sources
