+++ lua/lazyvim/util/init.lua
@@ -286,13 +286,4 @@ function M.memoize(fn)
   end
 end
 
----@return "nvim-cmp" | "blink.cmp"
-function M.cmp_engine()
-  vim.g.lazyvim_cmp = vim.g.lazyvim_cmp or "auto"
-  if vim.g.lazyvim_cmp == "auto" then
-    return LazyVim.has_extra("coding.nvim-cmp") and "nvim-cmp" or "blink.cmp"
-  end
-  return vim.g.lazyvim_cmp
-end
-
 return M
